Move watch progress to separate package.
